2025-05-09 CMS: 1.完成 Member - GetMemberDetail 功能跟單元測試 2.完成 Member - GetMemberDetail 前端 mock 跟 格式設計 3.前端元件 radioInput 版型調整，泛用性加強

+++ HealthopeCMS/ApiLayer/Models/Member/Response/ResponseGetMemberListDto.cs
@@ -52,7 +52,7 @@ namespace ApiLayer.Models.Member
         /// <summary>
         /// 今年團課缺席次數
         /// </summary>
-        public short AbsenceTime { get; set; }
+        public byte AbsenceTime { get; set; }
 
         /// <summary>
         /// 允許預約團課的時間(在那之前被禁止了)
